Small improvement

+++ tools/test.sh
@@ -16,7 +16,11 @@ if ! grep "All heap blocks were freed -- no leaks are possible" ./vg-test.log; t
   # The problem with 72,704 bytes can be ignored.
 
   if ! (grep "in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so" ./vg-test.log \
-        && grep "still reachable: 72,704 bytes in 1 blocks" ./vg-test.log); then
+        && grep "definitely lost: 0 bytes in 0 blocks" ./vg-test.log 
+        && grep "indirectly lost: 0 bytes in 0 blocks" ./vg-test.log 
+        && grep "possibly lost: 0 bytes in 0 blocks" ./vg-test.log 
+        && grep "still reachable: 72,704 bytes in 1 blocks" ./vg-test.log  
+        \); then
      rm ./vg-test.log
      echo "Memory leak error!"
      exit 1
